SoundS2CPayload#pitch changed to float
(cherry picked from commit ed9abacabf7f1eef840819905c3488fbc31ed2ce)

+++ src/main/java/io/github/lau6l/chatdiag/network/SoundS2CPayload.java
@@ -7,13 +7,13 @@ import net.minecraft.network.codec.PacketCodecs;
 import net.minecraft.network.packet.CustomPayload;
 import net.minecraft.util.Identifier;
 
-public record SoundS2CPayload(String soundId, int pitch) implements CustomPayload {
+public record SoundS2CPayload(String soundId, float pitch) implements CustomPayload {
     public static final Identifier SOUND_PAYLOAD_ID = ChatDiag.of("sound");
     public static final CustomPayload.Id<SoundS2CPayload> ID = new Id<>(SOUND_PAYLOAD_ID);
 
     public static final PacketCodec<RegistryByteBuf, SoundS2CPayload> CODEC = PacketCodec.tuple(
             PacketCodecs.STRING, SoundS2CPayload::soundId,
-            PacketCodecs.INTEGER, SoundS2CPayload::pitch,
+            PacketCodecs.FLOAT, SoundS2CPayload::pitch,
             SoundS2CPayload::new
     );
